move signer selection request thing to SignHandler
