+++ packages/wallet-sdk/src/sign/SignHandler.ts
@@ -3,7 +3,7 @@ import { SCWSigner } from './scw/SCWSigner';
 import { WLSigner } from './walletlink/WLSigner';
 import { PopUpCommunicator } from ':core/communicator/PopUpCommunicator';
 import { CB_KEYS_URL } from ':core/constants';
-import { SignerType } from ':core/message';
+import { ConfigMessage, SignerType } from ':core/message';
 import {
   AppMetadata,
   ConstructorOptions,
@@ -36,7 +36,7 @@ export class SignHandler {
   }
 
   async handshake() {
-    const signerType = await this.popupCommunicator.requestSignerSelection(this.preference);
+    const signerType = await this.requestSignerSelection();
     const signer = this.initSigner(signerType);
     const accounts = await signer.handshake();
 
@@ -64,6 +64,18 @@ export class SignHandler {
     return signerType ? this.initSigner(signerType) : null;
   }
 
+  private async requestSignerSelection(): Promise<SignerType> {
+    this.listenForWalletLinkSessionRequest();
+
+    const request: ConfigMessage = {
+      id: crypto.randomUUID(),
+      event: 'selectSignerType',
+      data: this.preference,
+    };
+    const { data } = await this.popupCommunicator.postMessage(request);
+    return data as SignerType;
+  }
+
   private initSigner(signerType: SignerType): Signer {
     const SignerClasses = {
       scw: SCWSigner,
@@ -75,5 +87,29 @@ export class SignHandler {
       popupCommunicator: this.popupCommunicator,
       updateListener: this.listener,
     });
+  }
+
+  // temporary walletlink signer instance to handle WalletLinkSessionRequest
+  // will revisit this when refactoring the walletlink signer
+  private listenForWalletLinkSessionRequest() {
+    this.popupCommunicator
+      .onMessage<ConfigMessage>(({ event }) => event === 'WalletLinkSessionRequest')
+      .then(async () => {
+        const walletlink = new WLSigner({
+          metadata: this.metadata,
+        });
+        this.postWalletLinkUpdate({ session: walletlink.getWalletLinkSession() });
+        // Wait for the wallet link session to be established
+        await walletlink.handshake();
+        this.postWalletLinkUpdate({ connected: true });
+      });
+  }
+
+  private postWalletLinkUpdate(data: unknown) {
+    const update: ConfigMessage = {
+      event: 'WalletLinkUpdate',
+      data,
+    };
+    this.popupCommunicator.postMessage(update);
   }
 }
